requirement.txt
''' A requirement list of the required modules needed to run this project '''
google-adk
google-generativeai
